+++ fe-mock/webpack.config.js
@@ -1,0 +1,77 @@
+const webpack = require('webpack');
+
+const port = process.env.PORT || 3000;
+
+module.exports = {
+    entry: "./src/index.js",
+    mode: 'development',
+    output: {
+        filename: '../../target/classes/static/built/bundle.js'
+    },
+    devtool: 'inline-source-map',
+    devServer: {
+        historyApiFallback: true,
+    },
+    module: {
+        rules: [
+            {
+                test: /\.(js)$/,
+                exclude: /node_modules/,
+                use: ['babel-loader']
+            },
+            {
+                test: /\.scss$/,
+                use: [
+                    {
+                        loader: 'style-loader'
+                    },
+                    {
+                        loader: 'css-loader',
+                        options: {
+                            sourceMap: true
+                        }
+                    },
+                    {
+                        // Run postcss actions
+                        loader: 'postcss-loader',
+                        options: {
+                            // `postcssOptions` is needed for postcss 8.x;
+                            // if you use postcss 7.x skip the key
+                            postcssOptions: {
+                                // postcss plugins, can be exported to postcss.config.js
+                                plugins: function () {
+                                    return [
+                                        require('autoprefixer')
+                                    ];
+                                }
+                            }
+                        }
+                    },
+                    {
+                        // compiles Sass to CSS
+                        loader: 'sass-loader',
+                        options: {
+                            sourceMap: true
+                        }
+                    },
+                ]
+            },
+            {
+                test: /\.svg$/,
+                use: [
+                    {
+                        loader: 'svg-url-loader',
+                        options: {
+                            limit: 10000,
+                        }
+                    }
+                ]
+            }
+        ]
+    },
+    plugins: [
+        new webpack.ProvidePlugin({
+            "React": "react",
+        }),
+    ]
+};