+++ src/main/java/eu/de4a/connector/mock/exampledata/DataOwner.java
@@ -8,7 +8,12 @@ import java.util.Arrays;
 public enum DataOwner {
     COC_NL("iso6523-actorid-upis::9999:NL990000106", "(KVK) Chamber of Commerce of Netherlands", "NL", Pilot.T42),
     V_SE("iso6523-actorid-upis::9999:SE000000013", "(BVE) BOLAGSVERKET (Companies Registration Office)", "SE", Pilot.T42),
-    ONRC_RO("iso6523-actorid-upis::9999:RO000000006", "(ORNC) - OFICIUL NATIONAL AL REGISTRULUI COMERTULUI", "RO", Pilot.T42);
+    ONRC_RO("iso6523-actorid-upis::9999:RO000000006", "(ORNC) - OFICIUL NATIONAL AL REGISTRULUI COMERTULUI", "RO", Pilot.T42),
+    DMDW_AT("iso6523-actorid-upis::9999:AT000000271", "(BMDW) Bundesministerium Fuer Digitalisierung Und Wirtschaftsstandort", "AT", Pilot.T42),
+    UL_PT("iso6523-actorid-upis::9999:PT990000101", "Portuguese IST, University of Lisbon", "PT", Pilot.T41),
+    JSI_SI("iso6523-actorid-upis::9999:SI000000018", "(JSI) Institut Jozef Stefan", "SI", Pilot.T41),
+    MIZS_SI("iso6523-actorid-upis::9999:SI000000016", "(MIZS) Ministrstvo za Izobrazevanje, Znanost in Sport", "SI", Pilot.T41),
+    UJI_ES("iso6523-actorid-upis::9999:ESQ6250003H", "(UJI) Universitat Jaume I de Castellón", "ES", Pilot.T41);
 
     @Getter
     final private String urn;
@@ -28,7 +33,7 @@ public enum DataOwner {
 
     public static DataOwner selectDataOwner(AgentType dataOwnerType) {
         return Arrays.stream(DataOwner.values())
-                .filter(dataOwner -> dataOwnerType.getAgentUrn().equals(dataOwner.getUrn()))
+                .filter(dataOwner -> dataOwnerType.getAgentUrn().equalsIgnoreCase(dataOwner.getUrn()))
                 .findFirst()
                 .orElseGet(() -> null);
     }